requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
ezdxf>=1.0.0
matplotlib>=3.5.0
numpy>=1.24.0
Pillow>=10.0.0
redis>=4.5.0
celery>=5.3.0
fastapi>=0.100.0
uvicorn>=0.23.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
python-jose>=3.3.0
passlib>=1.7.4
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2>=3.0.0 